Add Adaptive Bed Mesh Klipper module
This commit adds an new Klipper module for adaptive bed meshing.
The module relies on the exclude object detection to determine
the area of the bed being used by the print and only generate a
mesh for that area.

There are other macro-based solutions for this and there is also
an effort to get adaptive bed mesh included in mainline Klipper.
This is yet another implementation.

Note that until this feature is included in the master branch it
is considered experimental.